requirements.txt
pandas
requests
apache-airflow==2.9.1
pyarrow
pytest